refactor: canonize reputation contracts
