+++ apps/web/src/app/api/auth/twitter/callback/route.ts
@@ -40,7 +40,7 @@
  *           Location:
  *             schema:
  *               type: string
- *               example: /rewards?success=twitter_linked&points=100
+ *               example: /rewards?success=twitter_linked&reputation=100
  *       400:
  *         description: Invalid parameters or state expired
  *
@@ -50,14 +50,14 @@
  * // /api/auth/twitter/callback?code=abc123&state=user-id|timestamp|nonce
  *
  * // On success, user redirected to:
- * // /rewards?success=twitter_linked&points=100
+ * // /rewards?success=twitter_linked&reputation=100
  * ```
  *
  * @see {@link /api/auth/twitter/initiate} OAuth initiation
  * @see {@link /lib/services/points-service} Points service
  */
 
-import { PointsService, withErrorHandling } from '@babylon/api';
+import { ReputationService, withErrorHandling } from '@babylon/api';
 import { db } from '@babylon/db';
 import { getWaitlistBaseUrl, logger } from '@babylon/shared';
 import type { NextRequest } from 'next/server';
@@ -312,14 +312,14 @@ export const GET = withErrorHandling(async (request: NextRequest) => {
   });
 
   // Award points if this is the first time linking Twitter
-  const pointsResult = await PointsService.awardTwitterLink(
+  const pointsResult = await ReputationService.awardTwitterLink(
     userId,
     twitterUsername
   );
 
   // Check if this qualifies a referral (award bonus to referrer)
   if (pointsResult.success) {
-    await PointsService.checkAndQualifyReferral(userId).catch((error) => {
+    await ReputationService.checkAndQualifyReferral(userId).catch((error) => {
       // Log error but don't fail the request if qualification check fails
       logger.warn(
         `Failed to check and qualify referral for user ${userId}`,
@@ -331,14 +331,18 @@ export const GET = withErrorHandling(async (request: NextRequest) => {
 
   logger.info(
     'Twitter account linked successfully',
-    { userId, twitterUsername, pointsAwarded: pointsResult.pointsAwarded },
+    {
+      userId,
+      twitterUsername,
+      reputationAwarded: pointsResult.reputationAwarded,
+    },
     'TwitterCallback'
   );
 
   // Redirect back to configured destination with success
   return NextResponse.redirect(
     new URL(
-      `${OAUTH_REDIRECT_PATH}?success=twitter_linked&points=${pointsResult.pointsAwarded}`,
+      `${OAUTH_REDIRECT_PATH}?success=twitter_linked&reputation=${pointsResult.reputationAwarded}`,
       baseUrl
     )
   );
