+++ apps/web/src/app/api/users/[userId]/referrals/route.ts
@@ -75,6 +75,7 @@ import {
   follows,
   gte,
   inArray,
+  pointsTransactions,
   referrals,
   sum,
   tradingFees,
@@ -241,6 +242,22 @@ export const GET = withErrorHandling(
       .from(tradingFees)
       .where(eq(tradingFees.referrerId, canonicalUserId));
 
+    const [referralReputation] = await db
+      .select({
+        total: sum(pointsTransactions.amount),
+      })
+      .from(pointsTransactions)
+      .where(
+        and(
+          eq(pointsTransactions.userId, canonicalUserId),
+          inArray(pointsTransactions.reason, [
+            'referral_signup',
+            'referral_qualified',
+            'referral_bonus',
+          ])
+        )
+      );
+
     const totalFeesEarned = Number(feeEarnings?.total || 0);
 
     // Calculate weekly referral count (last 7 days) - only completed
@@ -358,6 +375,7 @@ export const GET = withErrorHandling(
       stats: {
         totalReferrals: completedReferralsData.length, // Only completed count
         pendingReferrals: pendingReferredUsers.length, // NEW: Pending count
+        totalReputationEarned: Number(referralReputation?.total || 0),
         totalFeesEarned,
         feeShareRate: 0.5, // 50% of fees
         followingCount: followingUserIds.size,
