+++ packages/api/src/services/onchain-service.ts
@@ -61,29 +61,29 @@ function resolveViemChain(chainId: number): Chain {
 }
 
 import { notifyNewAccount } from './notification-service';
-import { PointsService } from './points-service';
 import { getOrCreateReferralCode } from './referral-service';
+import { ReputationService } from './reputation-service';
 
 type OnboardingServices = {
   notifyNewAccount: (userId: string) => Promise<void>;
-  pointsService: {
+  reputationService: {
     awardReferralSignup: (
       referrerId: string,
       referredUserId: string
     ) => Promise<{
       success: boolean;
-      pointsAwarded: number;
+      reputationAwarded: number;
       error?: string;
     }>;
-    awardPoints: (
+    awardReputation: (
       userId: string,
       amount: number,
       reason: PointsReason,
       metadata?: StringRecord<JsonValue>
     ) => Promise<{
       success: boolean;
-      pointsAwarded: number;
-      newTotal: number;
+      reputationAwarded: number;
+      newReputationTotal: number;
     }>;
   };
   getOrCreateReferralCode: (userId: string) => Promise<string>;
@@ -112,9 +112,9 @@ function getOnboardingServices(): OnboardingServices {
 
   const fallback: OnboardingServices = {
     notifyNewAccount,
-    pointsService: {
-      awardReferralSignup: PointsService.awardReferralSignup,
-      awardPoints: PointsService.awardPoints,
+    reputationService: {
+      awardReferralSignup: ReputationService.awardReferralSignup,
+      awardReputation: ReputationService.awardReputation,
     },
     getOrCreateReferralCode,
   };
@@ -142,7 +142,7 @@ export interface OnchainRegistrationResult {
   message: string;
   tokenId?: number;
   txHash?: string;
-  pointsAwarded?: number;
+  reputationAwarded?: number;
   alreadyRegistered: boolean;
   userId: string;
 }
@@ -572,13 +572,13 @@ export async function processOnchainRegistration({
 
   // Process referrals
   if (referrerId) {
-    const referralResult = await services.pointsService.awardReferralSignup(
+    const referralResult = await services.reputationService.awardReferralSignup(
       referrerId,
       dbUser.id
     );
 
     if (referralResult.success) {
-      const refereeBonus = await services.pointsService.awardPoints(
+      const refereeBonus = await services.reputationService.awardReputation(
         dbUser.id,
         POINTS.REFERRAL_BONUS,
         'referral_bonus',
@@ -640,8 +640,8 @@ export async function processOnchainRegistration({
         {
           referrerId,
           referredUserId: dbUser.id,
-          referrerPoints: referralResult.pointsAwarded,
-          refereeBonus: refereeBonus.pointsAwarded,
+          referrerReputation: referralResult.reputationAwarded,
+          refereeReputationBonus: refereeBonus.reputationAwarded,
         },
         'OnboardingOnchain'
       );
@@ -688,7 +688,7 @@ export async function processOnchainRegistration({
     tokenId: agent0TokenId,
     txHash: registrationTxHash ?? undefined,
     alreadyRegistered: false,
-    pointsAwarded: 0,
+    reputationAwarded: 0,
     userId: dbUser.id,
   };
 }
