+++ packages/testing/unit/leaderboard-route.test.ts
@@ -69,6 +69,11 @@ mock.module('@babylon/api', () => ({
   findUserByIdentifier: mockFindUserByIdentifier,
   optionalAuth: mockOptionalAuth,
   getCache: mockGetCache,
+  ReputationService: {
+    getWalletLeaderboard: mockGetWalletLeaderboard,
+    getTeamLeaderboard: mockGetTeamLeaderboard,
+    getUserPosition: mockGetUserPosition,
+  },
   PointsService: {
     getWalletLeaderboard: mockGetWalletLeaderboard,
     getTeamLeaderboard: mockGetTeamLeaderboard,
